+++ lib/shipcloud/shipment.rb
@@ -5,7 +5,8 @@ module Shipcloud
     include Shipcloud::Operations::All
 
     attr_accessor :from, :to, :carrier, :package, :reference_number, :metadata
-    attr_reader :id, :created_at, :carrier_tracking_no, :tracking_url, :label_url, :packages, :price
+    attr_reader :id, :created_at, :carrier_tracking_no, :tracking_url, :label_url,
+                :packages, :price, :customs_declaration
 
     def self.index_response_root
       "#{class_name.downcase}s"