+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
 ### Added
 - Add attr_reader for `id` to class `Shipcloud::Address` to be able to get the id of a created address
-- Add attr_reader for `id` to class `Shipcloud::Webhook` to be able to get the id of a created webhook
+- Add attr_reader for `id` to class `Shipcloud::Webhook` to be able to get the id of the created webhook
 
 ### Changed